Move Screen capture from MainActivity to BrowserFragment
